+++ ios/Runner/Runner-Bridging-Header.h
@@ -1,3 +1,2 @@
 #import "GeneratedPluginRegistrant.h"
 #import <Irmagobridge/Irmagobridge.h>
-#import "IrmaMobileBridgePlugin.h"